requirements.txt
requests
boto3
dearpygui
Pillow
boto3
python-dotenv
rasterio